requirements.txt
azure-functions
tensorflow==1.13.1
Pillow
numpy
python-telegram-bot
requests